+++ base/src/main/java/io/github/sds100/keymapper/base/utils/ErrorUtils.kt
@@ -461,6 +461,9 @@ fun KMError.getFullMessage(resourceProvider: ResourceProvider): String {
         KMError.UnknownIOError ->
             resourceProvider.getString(R.string.error_io_error)
 
+        KMError.StarterFilesUnavailable ->
+            resourceProvider.getString(R.string.error_starter_files_unavailable)
+
         KMError.ShizukuNotStarted ->
             resourceProvider.getString(R.string.error_shizuku_not_started)